WIP: Restructure how verification requests are handled
A main goal here is that the vir-viper translation happens on the same
thread that encodes the program in the local verification case. This
should happen so the allocated arena, which is thread local, can be
used.
Additionally, the JVM reliant bits are a bit more hidden from the
general procedure.